+++ internal/server/remote.go
@@ -425,6 +425,8 @@ func (r *remotePlayer) emitEv(ev any, send func(byte, any)) {
 		send(attachproto.MsgShelfItems, ev)
 	case attachproto.MovingPiston:
 		send(attachproto.MsgMovingPiston, ev)
+	case attachproto.EntityAttributes:
+		send(attachproto.MsgEntityAttributes, ev)
 	case attachproto.WorldBorder:
 		send(attachproto.MsgWorldBorder, ev)
 	case attachproto.BannerPatterns: